start cleaning up CsharpMethod
