+++ src/CodeGeneration/ApiGenerator/Domain/Code/RequestParameterImplementation.cs
@@ -1,0 +1,12 @@
+using System.Collections.Generic;
+
+namespace ApiGenerator.Domain 
+{
+	public class RequestParameterImplementation
+	{
+		public CsharpNames CsharpNames { get; set; }
+		public string OfficialDocumentationLink { get; set; }
+		public IReadOnlyCollection<QueryParameters> Params { get; set; }
+		public string HttpMethod { get; set; }
+	}
+}
