+++ src/CodeGeneration/ApiGenerator/Views/IElasticLowLevelClient.Generated.cshtml
@@ -25,7 +25,7 @@ namespace Elasticsearch.Net
 			var endpoint = kv.Value;
 			foreach(var method in endpoint.GetCsharpMethods())
 			{
-<text>		///<summary>@method.HttpMethod on @method.Path <para>@method.DocumentationUrl</para></summary>
+<text>		///<summary>@method.HttpMethod on @method.Path <para>@method.OfficialDocumentationLink</para></summary>
 </text>
 		foreach (var part in method.Parts)
 		{
@@ -33,7 +33,7 @@ namespace Elasticsearch.Net
 </text>
 		}
 <text>		@Raw(@"///<param name=""requestParameters"">A func that allows you to describe the querystring parameters &amp; request specific connection settings.</param>")
-		@Raw(method.ReturnType) @(method.FullName)@(Raw(method.ReturnTypeGeneric))(@Raw(method.Arguments)) where @method.CallTypeGeneric : class, IElasticsearchResponse, new();
+		@Raw(method.ReturnType) @(method.PerPathMethodName)@(Raw("<TResponse>"))(@Raw(method.Arguments)) where TResponse : class, IElasticsearchResponse, new();
 </text>
 			}
 		}
